Model selection for BRIER.FULL fits

Selects the optimal eta combination and lambda value from a fitted BRIERfull object using held-out validation data. Information criteria (AIC, BIC, etc.) are not supported for BRIERfull because external observations are integrated via weights, which complicates the effective degrees-of-freedom adjustment.

Usage

BRIERfull.selection(
  object,
  criteria = c("gaussian.mspe", "gaussian.rsq", "binomial.dev", "binomial.mcfrsq",
    "binomial.tjursq", "binomial.auc", "poisson.dev"),
  X.val = NULL,
  y.val = NULL
)

Arguments

object

An object of class "BRIER" from BRIERfull.

criteria

A string specifying the validation criterion. One of:

"gaussian.mspe": mean squared prediction error.

"gaussian.rsq": negative R-squared (minimised).

"binomial.dev": binomial deviance.

"binomial.mcfrsq": McFadden's R-squared (negated, minimised).

"binomial.tjursq": Tjur's R-squared (negated, minimised).

"binomial.auc": AUC (negated, minimised).

"poisson.dev": Poisson deviance.

X.val

A numeric matrix of validation predictors (n_val x p).

y.val

A numeric response vector for validation (length n_val).

Value

An object of class c("BRIER.selection", "BRIER") which extends the input object with the following added elements:

criteria

The selection criterion used.

eta.min

The eta combination minimising the criterion.

eta.min.index

The corresponding row index in eta.grid.

lambda.min

The lambda value minimising the criterion at eta.min.

lambda.min.index

The corresponding lambda index.

eta.lambda

A data.frame summarising the optimal lambda for each eta.

See also

Examples

if (FALSE) { # \dontrun{
fit <- BRIERfull(
  X, y, cohort, family = "gaussian",
  eta.list = list(c(0, 0.5, 1, 2)),
  penalty = "LASSO"
)
sel <- BRIERfull.selection(
  fit, criteria = "gaussian.mspe",
  X.val = X.val, y.val = y.val
)
coef(sel)
} # }